Make interface equivalent to JSON.parse and JSON.stringify as an experiment
