--- test/fallback.properties
@@ -1,2 +1,0 @@
-hello=my ragtime gal
-another=question
