--- test/json/locales/us-EN/arrayTest.json
@@ -1,4 +1,0 @@
-{
-  "arrayTest": ["value1","value2","value3"],
-  "arrayTest2": ["value1","value2","value3"]
-}
