--- test/json/locales/us-EN/keyValueTest.json
@@ -1,6 +1,0 @@
-{
-  "keyValueTest1": "My Value 1",
-  "keyValueTest2": " My Value 2",
-  "keyValueTest3": "My Value 3 ",
-  "keyValueTest4": " My Value 4 "
-}
