--- test/json/locales/us-EN/mapTest.json
@@ -1,24 +1,0 @@
-{
-  "objectTest1": {
-    "key1": "My Value 1",
-    "key2": "My Value 2"
-  },
-  "objectTest2": {
-    "key1": "My Value 1",
-    "key2": {
-      "key1": "My Value 2"
-    },
-    "key3": {
-      "key2": {
-        "key1": "My Value 3"
-      }
-    },
-    "key4": {
-      "key3": {
-        "key2": {
-          "key1": "My Value 4"
-        }
-      }
-    }
-  }
-}
