--- test/main.properties
@@ -1,2 +1,0 @@
-hello=world
-include=./fallback.properties
