--- test/testData.txt
@@ -1,1 +1,0 @@
-key=value
